refactor: move volume thumbnailer code
