+++ src/core/thumbnailers/volume-thumbnailer.ts
